stopped permeture results and limited input
Before we made arrangement 100 times for all N. It works for small numbers not like 2, 3, 4. Not for 10, 12 etc. We ditched number of times method and implemented precision method. Which waits until every point is place at correct place with precision of 1e-10. For this to work we need to limit our input to 20. Yeah, QUALITY over QUANTITY.